Struct craft::summary::Summary

pub struct Summary { /* fields omitted */ }

Subset of a Manifest. Contains only the most important informations about a package. Summaries are cloned, and should not be mutated after creation

Methods

impl Summary

fn new(pkg_id: PackageId,
dependencies: Vec<Dependency>,
features: HashMap<String, Vec<String>>)
-> CraftResult<Summary>

fn package_id(&self) -> &PackageId

fn name(&self) -> &str

fn version(&self) -> &Version

fn source_id(&self) -> &SourceId

fn dependencies(&self) -> &[Dependency]

fn features(&self) -> &HashMap<String, Vec<String>>

fn checksum(&self) -> Option<&str>

fn override_id(self, id: PackageId) -> Summary

fn set_checksum(self, cksum: String) -> Summary

fn map_dependencies<F>(self, f: F) -> Summary where F: FnMut(Dependency) -> Dependency

fn map_source(self, to_replace: &SourceId, replace_with: &SourceId) -> Summary